# face_recognition_system/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Face Recognition System API",
    "description": "\n        ## 人脸识别系统 API\n\n        这是一个完整的人脸识别和门禁控制系统的RESTful API。\n\n        ### 主要功能\n        - **人员管理**: 添加、更新、删除和搜索人员信息\n        - **人脸特征管理**: 上传人脸图像，提取和管理人脸特征\n        - **实时人脸识别**: 识别和验证人脸身份\n        - **门禁控制**: 基于人脸识别的访问控制\n        - **统计报告**: 访问日志分析和统计报告\n        - **用户管理**: 系统用户和权限管理\n        - **系统管理**: 系统配置、备份和监控\n\n        ### 认证方式\n        API使用JWT (JSON Web Token) 进行认证。获取令牌后，在请求头中包含：\n        ```\n        Authorization: Bearer <your_access_token>\n        ```\n\n        ### 权限系统\n        - **read**: 读取权限 - 可以查看数据\n        - **write**: 写入权限 - 可以创建和修改数据\n        - **admin**: 管理员权限 - 可以执行所有操作\n\n        ### 速率限制\n        - 匿名用户: 100 请求/分钟\n        - 认证用户: 200 请求/分钟\n        - 管理员: 500 请求/分钟\n\n        ### 默认账户\n        - 管理员: `admin` / `admin123`\n        - 普通用户: `user` / `user123`\n        ",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "开发环境"
    },
    {
      "url": "https://api.example.com",
      "description": "生产环境"
    }
  ],
  "paths": {
...
    "/api/v1/users/": {
      "get": {
        "tags": [
          "user-management"
        ],
        "summary": "List Users",
        "description": "List all users.\nOnly admin users can access this endpoint.",
        "operationId": "list_users_api_v1_users__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserInfo"
                  },
                  "title": "Response List Users Api V1 Users  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "user-management"
        ],
        "summary": "Create User",
        "description": "Create a new user.\nOnly admin users can create new users.",
        "operationId": "create_user_api_v1_users__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Username"
            }
          },
          {
            "name": "password",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Password"
            }
          },
          {
            "name": "email",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Email"
            }
          },
          {
            "name": "full_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Full Name"
            }
          },
          {
            "name": "permissions",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "default": [
                "read"
              ],
              "title": "Permissions"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/persons/": {
      "get": {
        "tags": [
          "persons"
        ],
        "summary": "List Persons",
        "description": "List persons with optional filtering and pagination.\nReturns a paginated list of persons with optional filters.",
        "operationId": "list_persons_api_v1_persons__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sort By"
            }
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "asc",
              "title": "Sort Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "persons"
        ],
        "summary": "Create Person",
        "description": "Create a new person.\nCreates a new person record with the provided information.",
        "operationId": "create_person_api_v1_persons__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PersonCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Name"
          },
          "employee_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Employee Id"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "department": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Department"
          },
          "position": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "access_level": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 0.0,
            "title": "Access Level",
            "default": 1
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "PersonCreate",
        "description": "Person creation request model."
      },
      "PersonListResponse": {
        "properties": {
          "persons": {
            "items": {
              "$ref": "#/components/schemas/PersonResponse"
            },
            "type": "array",
            "title": "Persons"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          },
          "returned_count": {
            "type": "integer",
            "title": "Returned Count"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          }
        },
        "type": "object",
        "required": [
          "persons",
          "total_count",
          "returned_count",
          "offset",
          "limit",
          "has_more"
        ],
        "title": "PersonListResponse",
        "description": "Person list response model."
      },
      "PersonResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "uuid": {
            "type": "string",
            "title": "Uuid"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "employee_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Employee Id"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "department": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department"
          },
          "position": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "access_level": {
            "type": "integer",
            "title": "Access Level"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "face_feature_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Face Feature Count"
          },
          "last_access": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Access"
          }
        },
        "type": "object",
        "required": [
          "id",
          "uuid",
          "name",
          "employee_id",
          "email",
          "phone",
          "department",
          "position",
          "access_level",
          "is_active",
          "notes",
          "created_at",
          "updated_at"
        ],
        "title": "PersonResponse",
        "description": "Person response model."
      },
...
      "UserInfo": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Permissions"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "last_login": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "email",
          "full_name",
          "is_active",
          "permissions",
          "created_at",
          "last_login"
        ],
        "title": "UserInfo",
        "description": "User information model."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT认证令牌"
      }
    },
...
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
}